paper-to-slides/requirements.txt
# Core dependencies
PyMuPDF==1.24.2          # fitz for PDF processing
pdfplumber==0.10.3
pillow==10.3.0           # PIL for image processing
opencv-python==4.9.0.80
numpy==1.26.4
anthropic==0.25.4        # Claude integration
openai==1.30.1
jinja2==3.1.3
pyyaml==6.0.1
beautifulsoup4==4.12.3   # HTML parsing for accessibility
python-dotenv==1.0.1     # Environment variable management

# Development & testing
pytest==8.1.1
pytest-mock==3.14.0
python-slugify==8.0.4

# Optional for enhanced figure processing
scikit-image==0.22.0     # For image enhancement
svglib==1.5.1            # SVG conversion
reportlab==4.1.0         # PDF generation (for svglib) 